+++ lib/redis_failover.rb
@@ -1,4 +1,3 @@
-require 'set'
 require 'redis'
 require 'thread'
 require 'logger'