+++ Changes.md
@@ -1,7 +1,7 @@
 HEAD
 -----------
 - No longer force newly available slaves to master if already slaves of that master
-- Honor a node's slave-serve-stale-data configuration option; do not mark a sync-with-master-in-progress slave as available if its slave-serve-stale-data is disabled)
+- Honor a node's slave-serve-stale-data configuration option; do not mark a sync-with-master-in-progress slave as available if its slave-serve-stale-data is disabled
 - Change reachable/unreachable wording to available/unavailable
 
 0.3.0