+++ Changes.md
@@ -1,4 +1,4 @@
-HEAD
+0.4.0
 -----------
 - No longer force newly available slaves to master if already slaves of that master
 - Honor a node's slave-serve-stale-data configuration option; do not mark a sync-with-master-in-progress slave as available if its slave-serve-stale-data is disabled